+++ novel/src/main/java/com/wuming/novel/domain/entity/Job.java
@@ -12,8 +12,8 @@ import java.time.LocalDateTime;
 @TableName("jobs")
 public class Job implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Integer id;
-    private Integer novelId;
+    private Long id;
+    private Long novelId;
     private String protagonistName;      // 用户将会扮演的角色
     private String targetName;      // llm将会扮演的角色
     private JobStage stage;
