+++ novel/src/main/java/com/wuming/novel/serviceImpl/PipelineService.java
@@ -24,7 +24,7 @@ public class PipelineService {
 
     // TODO 将来各个阶段抛出异常则从原处恢复
     // TODO 针对前面几个与job无关的阶段可以检测是否已经完成，完成则跳过
-    public boolean handleNovel(int jobId) throws IOException {
+    public boolean handleNovel(Long jobId) throws IOException {
         Job job = jobService.getById(jobId);
         if(job == null){
             throw new IllegalArgumentException("该job不存在，请创建后重试");
